refactor: コード品質の改善（コンポーネント分割と重複削減）
## 実装内容

### 1. マジックナンバーの定数化

**frontend/src/hooks/useKeyboardShortcuts.ts**
- `setTimeout(..., 1000)` を `SEQUENCE_RESET_TIMEOUT_MS` 定数化
- コメントで意図を明確化（キーシーケンス入力のリセット時間）

### 2. 重複コードの削減

**frontend/src/pages/TaskList.tsx**
- ステータスフィルター切り替えロジックを `toggleStatusFilter()` 関数に統一
- 検索フォーカス処理を `focusSearchInput()` 関数に統一
- 約70行から40行に削減（約43%のコード削減）

**Before（重複あり）:**
```typescript
{
  key: "1",
  action: () => {
    const currentStatus = sp.getAll("status") as Status[];
    const next = new URLSearchParams(sp);
    const statusSet = new Set(currentStatus);
    // ... 15行の重複ロジック
  }
}
// 2番、3番も同様に重複
```

**After（統一化）:**
```typescript
const toggleStatusFilter = (targetStatus: Status) => {
  // ... 統一されたロジック
};

{
  key: "1",
  action: () => toggleStatusFilter("not_started"),
}
```

### 3. TaskDrawerコンポーネントの分割

**新規ファイル: frontend/src/features/drawer/TaskChildForm.tsx**
- 子タスク作成フォームを独立コンポーネント化
- 責務: 子タスクの入力、バリデーション、作成処理
- Props: `taskId`, `onSuccess` コールバック

**frontend/src/features/drawer/TaskDrawer.tsx**
- TaskChildForm を使用するように修正
- 約60行のコード削減
- useState の数を3個削減（childTitle, childDue, creating）
- 不要なimportを削除（useCreateTask, brandIso, getUserMessage, logError）

**効果:**
- TaskDrawerInner が約365行 → 305行に（約16%削減）
- 単一責任原則の適用
- テストしやすさの向上

## 改善効果

### ✅ 保守性の向上
- コンポーネントが小さく理解しやすい
- 責務が明確（TaskChildForm = 子タスク作成のみ）

### ✅ 可読性の向上
- マジックナンバーの排除
- 重複コードの統一化でDRY原則を適用

### ✅ 再利用性の向上
- TaskChildForm は他の場所でも使用可能
- toggleStatusFilter は明確なインターフェース

### ✅ テストしやすさの向上
- 小さなコンポーネントは単体テストが容易
- ヘルパー関数も独立してテスト可能

## テスト結果

✅ 全75テスト成功 - 既存機能に影響なし

## 影響範囲

**修正ファイル数:** 3ファイル
- hooks/useKeyboardShortcuts.ts
- pages/TaskList.tsx
- features/drawer/TaskDrawer.tsx

**新規ファイル数:** 1ファイル
- features/drawer/TaskChildForm.tsx

Closes #283
