+++ frontend/src/pages/TaskList.tsx
@@ -22,91 +22,66 @@ const TaskList: PageComponent = () => {
 
   const [sp, setSp] = useSearchParams();
 
+  /**
+   * ステータスフィルターを切り替えるヘルパー関数
+   * 指定されたステータスが既に選択されている場合は解除、
+   * 選択されていない場合は単独で選択する
+   */
+  const toggleStatusFilter = (targetStatus: Status) => {
+    const currentStatus = sp.getAll("status") as Status[];
+    const next = new URLSearchParams(sp);
+    const statusSet = new Set(currentStatus);
+
+    if (statusSet.has(targetStatus)) {
+      statusSet.delete(targetStatus);
+    } else {
+      statusSet.clear();
+      statusSet.add(targetStatus);
+    }
+
+    next.delete("status");
+    for (const s of statusSet) {
+      next.append("status", s);
+    }
+    setSp(next, { replace: true });
+  };
+
+  /**
+   * 検索ボックスにフォーカスを当てるヘルパー関数
+   */
+  const focusSearchInput = () => {
+    const searchInput = document.querySelector('[data-testid="search-input"]') as HTMLInputElement;
+    searchInput?.focus();
+  };
+
   // Keyboard shortcuts for TaskList page
   useKeyboardShortcuts([
     // Search focus (F or /)
     {
       key: "f",
       description: "検索ボックスにフォーカス",
-      action: () => {
-        const searchInput = document.querySelector('[data-testid="search-input"]') as HTMLInputElement;
-        searchInput?.focus();
-      },
+      action: focusSearchInput,
     },
     {
       key: "/",
       description: "検索ボックスにフォーカス",
-      action: () => {
-        const searchInput = document.querySelector('[data-testid="search-input"]') as HTMLInputElement;
-        searchInput?.focus();
-      },
+      action: focusSearchInput,
     },
     // Filter by status
     {
       key: "1",
       description: "未着手のみ表示",
-      action: () => {
-        const currentStatus = sp.getAll("status") as Status[];
-        const next = new URLSearchParams(sp);
-        const statusSet = new Set(currentStatus);
-
-        if (statusSet.has("not_started")) {
-          statusSet.delete("not_started");
-        } else {
-          statusSet.clear();
-          statusSet.add("not_started");
-        }
-
-        next.delete("status");
-        for (const s of statusSet) {
-          next.append("status", s);
-        }
-        setSp(next, { replace: true });
-      },
+      action: () => toggleStatusFilter("not_started"),
     },
     {
       key: "2",
       description: "進行中のみ表示",
-      action: () => {
-        const currentStatus = sp.getAll("status") as Status[];
-        const next = new URLSearchParams(sp);
-        const statusSet = new Set(currentStatus);
-
-        if (statusSet.has("in_progress")) {
-          statusSet.delete("in_progress");
-        } else {
-          statusSet.clear();
-          statusSet.add("in_progress");
-        }
-
-        next.delete("status");
-        for (const s of statusSet) {
-          next.append("status", s);
-        }
-        setSp(next, { replace: true });
-      },
+      action: () => toggleStatusFilter("in_progress"),
     },
     {
       key: "3",
       description: "完了のみ表示",
-      action: () => {
-        const currentStatus = sp.getAll("status") as Status[];
-        const next = new URLSearchParams(sp);
-        const statusSet = new Set(currentStatus);
-
-        if (statusSet.has("completed")) {
-          statusSet.delete("completed");
-        } else {
-          statusSet.clear();
-          statusSet.add("completed");
-        }
-
-        next.delete("status");
-        for (const s of statusSet) {
-          next.append("status", s);
-        }
-        setSp(next, { replace: true });
-      },
+      action: () => toggleStatusFilter("completed"),
     },
     {
       key: "0",
